pyproject.toml
[project]
name = "meridian-chatbot"
version = "0.1.0"
description = "AI-powered customer support chatbot for Meridian Electronics"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
    "mcp>=1.9.0",
    "gradio>=5.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
